requirements.txt
python-dotenv~=1.0.1
Flask~=3.0.3
requests~=2.32.3
pytest~=8.3.3